college_bank_backend/urls.py
from django.contrib import admin
from django.urls import path, include
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
)

urlpatterns = [
    path('admin/', admin.site.urls),
    path('api/', include('bank_accounts.urls')), # Include your app's URLs
   

    # JWT Authentication URLs
    path('api/auth/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/auth/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    # You might also want a token verify endpoint:
    # path('api/auth/token/verify/', TokenVerifyView.as_view(), name='token_verify'),
]